+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [1.7.7] - 2025-06-10
+
+### Added
+- The `describe file types` attachment modifier.
+
+### Fixed
+- Issue with retrieving the JSON representation of `question` that
+  uses `list collect` and `uncheck others`.
+
 ## [1.7.6] - 2025-05-23
 
 ### Added